# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/shops/{shopId}": {
      "get": {
        "operationId": "ShopsController_get",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El taller con mi rol.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopView"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no soy miembro (no se distingue)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalle de un taller (solo miembros).",
        "tags": [
          "shops"
        ]
      }
    },
...
    "/v1/shops/{shopId}/vehicles": {
      "post": {
        "operationId": "VehiclesController_create",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehículo creado (data_source: vin_decode | manual).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleView"
                }
              }
            }
          },
          "409": {
            "description": "VIN/chasis/placa ya registrado en el taller; trae existing_vehicle_id."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registra un vehículo. Con VIN se autocompleta; sin VIN, carga manual.",
        "tags": [
          "vehicles"
        ]
      },
      "get": {
        "operationId": "VehiclesController_search",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Placa, VIN, chasis (parciales) o marca/modelo.",
            "schema": {
              "example": "A123",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "next_cursor de la página anterior.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer_id",
            "required": false,
            "in": "query",
            "description": "Solo los vehículos de este cliente.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de vehículos, del más nuevo al más viejo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleViewPage"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Busca vehículos del taller (placa, VIN, chasis, marca/modelo).",
        "tags": [
          "vehicles"
        ]
      }
    },
    "/v1/shops/{shopId}/vehicles/{vehicleId}": {
      "get": {
        "operationId": "VehiclesController_get",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleView"
                }
              }
            }
          },
          "404": {
            "description": "No existe en este taller."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalle de un vehículo del taller.",
        "tags": [
          "vehicles"
        ]
      },
      "patch": {
        "operationId": "VehiclesController_update",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleView"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualiza un vehículo (kilometraje, placa, dueño…). null borra un dato.",
        "tags": [
          "vehicles"
        ]
      }
    },
    "/v1/shops/{shopId}/customers/{customerId}/vehicles": {
      "get": {
        "operationId": "VehiclesController_byCustomer",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Placa, VIN, chasis (parciales) o marca/modelo.",
            "schema": {
              "example": "A123",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "next_cursor de la página anterior.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleViewPage"
                }
              }
            }
          },
          "404": {
            "description": "El cliente no existe en este taller."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vehículos de un cliente del taller.",
        "tags": [
          "vehicles"
        ]
      }
    },
...
  },
  "info": {
    "title": "MechMate API",
    "description": "Backend del ecosistema MechMate (ERP de talleres + marketplace de repuestos, RD). Dinero en centavos como string (`Cents`); fechas ISO-8601.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "customer_id": {
            "type": "string",
            "description": "Cliente dueño (del mismo taller)."
          },
          "vin": {
            "type": "object",
            "example": "1HGCM82633A004352",
            "nullable": true,
            "description": "Si viene, autocompleta marca/modelo/año/motor."
          },
          "chassis_number": {
            "type": "object",
            "example": "NZE121-1234567",
            "nullable": true,
            "description": "Para vehículos sin VIN (p. ej. japoneses)."
          },
          "plate": {
            "type": "object",
            "example": "A123456",
            "nullable": true
          },
          "make": {
            "type": "string",
            "example": "Toyota",
            "description": "Obligatoria si no hay VIN decodificable."
          },
          "model": {
            "type": "object",
            "example": "Corolla",
            "nullable": true
          },
          "year": {
            "type": "object",
            "example": 2019,
            "nullable": true
          },
          "trim": {
            "type": "object",
            "nullable": true
          },
          "engine": {
            "type": "object",
            "example": "1.8L 4 cil.",
            "nullable": true
          },
          "fuel_type": {
            "type": "object",
            "example": "Gasolina",
            "nullable": true
          },
          "color": {
            "type": "object",
            "example": "Gris",
            "nullable": true
          },
          "mileage_km": {
            "type": "object",
            "example": 85000,
            "nullable": true
          },
          "notes": {
            "type": "object",
            "nullable": true,
            "maxLength": 1000
          }
        },
        "required": [
          "customer_id"
        ]
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {
          "customer_id": {
            "type": "string",
            "description": "Cliente dueño (del mismo taller)."
          },
          "vin": {
            "type": "object",
            "example": "1HGCM82633A004352",
            "nullable": true,
            "description": "Si viene, autocompleta marca/modelo/año/motor."
          },
          "chassis_number": {
            "type": "object",
            "example": "NZE121-1234567",
            "nullable": true,
            "description": "Para vehículos sin VIN (p. ej. japoneses)."
          },
          "plate": {
            "type": "object",
            "example": "A123456",
            "nullable": true
          },
          "make": {
            "type": "string",
            "example": "Toyota",
            "description": "Obligatoria si no hay VIN decodificable."
          },
          "model": {
            "type": "object",
            "example": "Corolla",
            "nullable": true
          },
          "year": {
            "type": "object",
            "example": 2019,
            "nullable": true
          },
          "trim": {
            "type": "object",
            "nullable": true
          },
          "engine": {
            "type": "object",
            "example": "1.8L 4 cil.",
            "nullable": true
          },
          "fuel_type": {
            "type": "object",
            "example": "Gasolina",
            "nullable": true
          },
          "color": {
            "type": "object",
            "example": "Gris",
            "nullable": true
          },
          "mileage_km": {
            "type": "object",
            "example": 85000,
            "nullable": true
          },
          "notes": {
            "type": "object",
            "nullable": true,
            "maxLength": 1000
          }
        }
      },
...
      "ShopView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ShopType"
          },
          "commission_bps": {
            "type": "number"
          },
          "my_role": {
            "$ref": "#/components/schemas/ShopMemberRole"
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "commission_bps",
          "my_role",
          "created_at"
        ],
        "additionalProperties": false
      },
      "ShopType": {
        "type": "string",
        "enum": [
          "mechanic_shop",
          "parts_seller"
        ]
      },
      "ShopMemberRole": {
        "type": "string",
        "enum": [
          "owner",
          "mechanic",
          "advisor"
        ]
      },
...
      "VehicleView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "customer_id": {
            "type": "string"
          },
          "vin": {
            "type": "string",
            "nullable": true
          },
          "chassis_number": {
            "type": "string",
            "nullable": true
          },
          "plate": {
            "type": "string",
            "nullable": true
          },
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "number",
            "nullable": true
          },
          "trim": {
            "type": "string",
            "nullable": true
          },
          "engine": {
            "type": "string",
            "nullable": true
          },
          "fuel_type": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "mileage_km": {
            "type": "number",
            "nullable": true
          },
          "data_source": {
            "type": "string",
            "enum": [
              "vin_decode",
              "manual"
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "customer_id",
          "vin",
          "chassis_number",
          "plate",
          "make",
          "model",
          "year",
          "trim",
          "engine",
          "fuel_type",
          "color",
          "mileage_km",
          "data_source",
          "notes",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
...
      "VehicleViewPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleView"
            }
          },
          "next_cursor": {
            "type": "string",
            "nullable": true,
            "description": "Pasar como `cursor` para la siguiente página; null si no hay más."
          }
        },
        "required": [
          "items",
          "next_cursor"
        ]
      },
...
    }
  }
}